Replaced several item names
This change will require users to "modify" their client, by copying a text file over to its folder, but without this, many items would just not work at all, because their names are too long for the char[16] array that item names are sent in. Some of the problematic items are known to have been present in the alpha, however, so we needed a solution. Alternatively, we could've sent the Korean names for these items, but it's preferrable to mod the alpha client anyway, and maybe we could simply host a complete mod package somewhere, that users just have to copy and paste.