Modified protocol versions slightly
Trying to use the protocol versions eAthena used early on doesn't work well, because there are minor differences in the packet tables even between versions that should presumably be on the same protocol version. They also started at version 5, but I've already identified more than 5 versions in the early clients. Dates wouldn't work well either though, because the server currently doesn't focus on one region, but uses whatever I can get my hands on, such as iRO, jRO, and euRO, and "version" 20030101 for example will presumably not be the same across multiple versions. So I'm switching to a custom system starting at 100, which I increase in steps of 100 for major versions to leave room in between.